requirements.txt
django
pillow
ipython
bpython